PLANET-4425 fix rendering of columns block with invalid attr
Use a hook that is called before the blocks attributes are validated.
Unfortunately the block renderer validates the request prior to when the
block is rendered, and that hook is not called there.

Furthermore there doesn't seem to be a hook in the REST API to modify
the block to be rendered before it is validated, so the only option was
the modify this on the frontend just before it calls the API.