WIP make custom fonts work again in editor
* For now this causes the titles in the editor sidebar to follow the
style too, which looks interesting but probably not what we want.
* It probably broke with a WP version update that changed the
DOMContentLoaded event the function relied on to add the event
listeners. It now gets called when the post type is not there yet.
* Instead now it checks this in the listener. Since it just needs to
read an in memory string, it can get called frequently.
* It was broken a second time (but not noticed since it already was
broken) because the variable names changed. Fixed these too.