PLANET-5872 Run a11y against test instances
Ref: https://jira.greenpeace.org/browse/PLANET-5872

---

Re-enables a11y tests, but instead of spawning a full docker-compose
setup, move the job a bit further in the pipeline and use the test
instance that's already deployed.

For now the job passes in any case. Once we refine a11y tickets and fix
the issues we can account for the errors and make it fail.